sandbox/cgroup: improve pid parsing code
Move all the pid parsing code to pids.go and arrange in logical
progression. Add extra tests to improve coverage.
